// src/@types/serverless/aws.d.ts
declare namespace Aws {
  /*
        Types based on https://github.com/serverless/serverless/blob/master/docs/providers/aws/guide/serverless.yml.md
    */
  interface Serverless {
    service: Service | string
    useDotenv?: boolean | undefined
    frameworkVersion?: string | undefined
    enableLocalInstallationFallback?: boolean | undefined
    variablesResolutionMode?: '20210219' | '20210326' | undefined
    unresolvedVariablesNotificationMode?: 'warn' | 'error' | undefined
    deprecationNotificationMode?: 'warn' | 'warn:summary' | 'error' | undefined
    disabledDeprecations?: string[] | undefined
    configValidationMode?: 'warn' | 'error' | 'off' | undefined
    provider: Provider
    package?: Package | undefined
    functions?: Functions | undefined
    layers?: Layers | undefined
    resources?: Resources | undefined
    plugins?: string[] | undefined
    org?: string | undefined
    app?: string | undefined
    tenant?: string | undefined
    custom?: Custom | undefined
  }

  interface Service {
    name: string
    /** @deprecated in favor of `kmsKeyArn` at the provider level  */
    awsKmsKeyArn?: string | undefined
  }

  interface Provider {
    name: 'aws'
    runtime?: string | undefined
    stage?: string | undefined
    region?: string | undefined
    stackName?: string | undefined
    apiName?: string | undefined
    lambdaHashingVersion?: number | undefined
    websocketsApiName?: string | undefined
    websocketsApiRouteSelectionExpression?: string | undefined
    profile?: string | undefined
    memorySize?: number | string | undefined
    reservedConcurrency?: number | string | undefined
    timeout?: number | string | undefined
    logRetentionInDays?: number | string | undefined
    deploymentBucket?: DeploymentBucket | undefined
    deploymentPrefix?: string | undefined
    /** @deprecated in favor of `iam.role` */
    role?: string | undefined
    /** @deprecated in favor of `iam.role.permissionsBoundary` */
    rolePermissionsBoundary?: string | undefined
    /** @deprecated in favor of `iam.role.statements` */
    iamRoleStatements?: IamRoleStatement[] | undefined
    /** @deprecated in favor of `iam.role.managedPolicies` */
    iamManagedPolicies?: string[] | undefined
    /** @deprecated in favor of `iam.deploymentRole` */
    cfnRole?: string | undefined
    iam?: IamSettings | undefined
    versionFunctions?: boolean | undefined
    architecture?: 'x86_64' | 'arm64' | undefined
    environment?: Environment | string | undefined
    endpointType?: 'regional' | 'edge' | 'private' | undefined
    apiKeys?: string[] | undefined
    apiGateway?: ApiGateway | undefined
    alb?: Alb | undefined
    httpApi?: HttpApi | undefined
    usagePlan?: UsagePlan | undefined
    stackTags?: Tags | undefined
    stackPolicy?: ResourcePolicy[] | undefined
    vpc?: string | Vpc | undefined
    notificationArns?: string[] | undefined
    stackParameters?: StackParameters[] | undefined
    resourcePolicy?: ResourcePolicy[] | undefined
    rollbackConfiguration?: RollbackConfiguration | undefined
    tags?: Tags | undefined
    tracing?: Tracing | undefined
    logs?: Logs | undefined
    kmsKeyArn?: string | undefined
    eventBridge?: EventBridge | undefined
    layers?: Array<string | Record<string, string>> | undefined
  }
// ...
  interface RestApiLogs {
    accessLogging?: boolean | undefined
    format?: string | undefined
    executionLogging?: boolean | undefined
    level?: 'INFO' | 'ERROR' | undefined
    fullExecutionData?: boolean | undefined
    role?: string | undefined
    roleManagedExternally?: boolean | undefined
  }

  interface WebsocketLogs {
    level?: 'INFO' | 'ERROR' | undefined
  }

  interface HttpApiLogs {
    format?: string | undefined
  }

  interface Logs {
    restApi?: true | RestApiLogs | undefined
    websocket?: WebsocketLogs | undefined
    httpApi?: boolean | HttpApiLogs | undefined
    frameworkLambda?: boolean | undefined
  }

  interface Package {
    /** @deprecated use `patterns` instead */
    include?: string[] | undefined
    /** @deprecated use `patterns` instead */
    exclude?: string[] | undefined
    patterns?: string[] | undefined
    excludeDevDependencies?: boolean | undefined
    artifact?: string | undefined
    individually?: boolean | undefined
  }
// ...
  interface Functions {
    [key: string]: AwsFunctionHandler | AwsFunctionImage
  }

  interface Layer {
    path: string
    name?: string | undefined
    description?: string | undefined
    compatibleRuntimes?: string[] | undefined
    licenseInfo?: string | undefined
    allowedAccounts?: string[] | undefined
    retain?: boolean | undefined
  }

  interface Layers {
    [key: string]: Layer
  }
// ...
  interface Resources {
    Description?: string | undefined
    Conditions?: ResourcesConditions
    Resources: CloudFormationResources
    extensions?: CloudFormationResources | undefined
    Outputs?: Outputs | undefined
  }

  interface Custom {
    [key: string]: any
  }

  interface Credentials {
    [key: string]: any
  }
}

declare class Aws {
  constructor(serverless: Serverless, options: Serverless.Options)

  naming: { [key: string]: () => string }
  getCredentials(): Aws.Credentials
  getProviderName(): string
  getRegion(): string
  getServerlessDeploymentBucketName(): Promise<string>
  getStage(): string
  getAccountId(): Promise<string>
  request(
    service: string,
    method: string,
    params?: {},
    options?: { useCache?: boolean | undefined; region?: string | undefined },
  ): Promise<any>
}